% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Function to sum}
\usage{
add(a, b)
}
\arguments{
\item{a}{This is the first parameter as numeric}

\item{b}{this is the second parameter as numeric}
}
\value{
sum of a and b
}
\description{
Function to sum
}
\examples{
add(2,5)
}
